+++ client/src/model/GlobalSchema.js
@@ -26,7 +26,7 @@
 import { Schema, PropertyName as Prop } from "./index";
 import {
   notebooksSchema, userSchema, projectsSchema, projectGlobalSchema, newProjectSchema,
-  statuspageSchema, notificationsSchema
+  statuspageSchema, notificationsSchema, formGeneratorSchema
 } from "./RenkuModels";
 
 const globalSchema = new Schema({
@@ -36,7 +36,8 @@ const globalSchema = new Schema({
   project: { [Prop.SCHEMA]: projectGlobalSchema },
   newProject: { [Prop.SCHEMA]: newProjectSchema },
   statuspage: { [Prop.SCHEMA]: statuspageSchema },
-  notifications: { [Prop.SCHEMA]: notificationsSchema }
+  notifications: { [Prop.SCHEMA]: notificationsSchema },
+  formGenerator: { [Prop.SCHEMA]: formGeneratorSchema }
 });
 
 export { globalSchema };