+++ client/src/utils/formgenerator/fields/CreatorsInput.js
@@ -123,11 +123,13 @@ function CreatorsInput({ name, label, type, value, alert, placeholder, setInputs
   useEffect(()=>{
     const filteredCreators = creators.filter(creator =>
       !(creator.name === "" && creator.email === "" && creator.affiliation === ""));
-    const artificialEvent = {
-      target: { name: name, value: filteredCreators },
-      isPersistent: () => false
-    };
-    setInputs(artificialEvent);
+    if (filteredCreators.length !== value.length) {
+      const artificialEvent = {
+        target: { name: name, value: filteredCreators },
+        isPersistent: () => false
+      };
+      setInputs(artificialEvent);
+    }
   // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [creators]);